test(hwpx): 구역 매퍼 공유 픽스처를 분리한다
HwpxSectionMapperTests가 SwiftLint type_body_length error 임계(400줄)에
닿아 새 테스트를 넣을 수 없었다. id 테이블·구역 래퍼·최소 구역 픽스처를
HwpxSectionFixture로 옮긴다 (헤더 스위트의 HwpxHeaderFixture와 같은
전례). 테스트 단언은 그대로다.
